Added logic to choose 'short' or 'scratch' as default shortpath location (#212)
* Added logic to choose 'short' or 'scratch' as default shortpath
location depending on which exists

* Added tests for lab path creation.

* Rearranged existing tests and shared code to common.py

* Fixed travis to run all tests including new one
